requirements.txt
flask
scikit-learn
pandas
numpy
Pillow
matplotlib
seaborn
joblib
